requirements.txt
# ============================================
# UNIFIED REQUIREMENTS - Scraper Backend
# ============================================
# Python 3.11+ recommended

# ==================== Core ====================
# Web scraping framework
scrapy==2.12.0
scrapy-zyte-api==0.31.0
itemadapter==0.10.0

# HTTP requests
requests==2.31.0

# ==================== Database ====================
supabase==2.24.0

# ==================== Data Processing ====================
pandas==2.2.3
beautifulsoup4==4.12.2

# ==================== Web Server ====================
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# ==================== Browser Automation ====================
# Selenium for local browser automation
selenium==4.15.2
webdriver-manager==4.0.1
# Zyte Smart Proxy Manager integration for Selenium - using direct proxy config (same as scrapers)
# zyte-smartproxy-selenium package not needed - using --proxy-server approach
# Playwright for Browserless.io fallback (when bot detection occurs) - deprecated, using Zyte instead
playwright==1.40.0
# Nodriver for CDP-free automation (bypasses bot detection better than Selenium)
# See: https://roundproxies.com/blog/bypass-bot-detection/
nodriver==0.41

# ==================== Utilities ====================
python-dotenv==1.0.0
schedule==1.2.2
